+++ src/main/java/com/kefa/api/dto/account/request/AccountLoginRequest.java
@@ -1,4 +1,4 @@
-package com.kefa.api.dto.request.account;
+package com.kefa.api.dto.account.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
