+++ src/main/java/com/kefa/api/dto/account/response/AccountUpdatePasswordResponse.java
@@ -1,4 +1,4 @@
-package com.kefa.api.dto.response.account;
+package com.kefa.api.dto.account.response;
 
 import lombok.*;
 
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 @Builder
-public class AccountUpdatePasswordResponseDto {
+public class AccountUpdatePasswordResponse {
 
     private final LocalDateTime updateAt = LocalDateTime.now();
     private final String message = "비밀번호 변경 완료";
