+++ src/main/java/com/kefa/api/dto/account/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.kefa.api.dto.response.account;
+package com.kefa.api.dto.account.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
